requirements.txt
anthropic>=0.34.0
openai>=1.40.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
moviepy>=1.0.3
pydub>=0.25.1
requests>=2.31.0
python-dotenv>=1.0.0
mutagen>=1.47.0
Pillow>=10.0.0
pytrends>=4.9.2
feedparser>=6.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
newsapi-python>=0.2.7
